pyproject.toml
[project]
name = "ai-servant-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "fastapi[standard]>=0.129.0",
    "pydantic-settings>=2.13.0",
    "sqlalchemy[asyncio]>=2.0.46",
]

[dependency-groups]
dev = [
    "alembic>=1.18.4",
    "black>=26.1.0",
    "ruff>=0.15.1",
]



[tool.ruff]
# Версия Python
target-version = "py312"

# Какие файлы проверять
include = ["src/*.py", "*.py"]
exclude = [
    ".venv",
    "dist",
    "build",
]

# Максимальная длина строки
line-length = 88

[tool.ruff.lint]
# Выбираем правила
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # Ignore line-too-long (black manages it)
]

[tool.ruff.format]
# Использовать как черный форматтер
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
